+++ src/Core/Application/Exchange/Offers/OfferDetailsDto.cs
@@ -6,10 +6,10 @@ public class OfferDetailsDto : IDto
 {
     public Guid Id { get; set; }
     public string CurrencyCode { get; set; } = default!;
-    public decimal NetValue { get; set; }
-    public decimal GrossValue { get; set; }
+    public long NetValue { get; set; }
+    public long GrossValue { get; set; }
     public DeliveryCostType DeliveryCostType { get; set; }
-    public decimal DeliveryCostGrossPrice { get; set; }
+    public long DeliveryCostGrossPrice { get; set; }
     public string? DeliveryCostDescription { get; set; }
     public DateOnly? ExpirationDate { get; set; }
     public string? Freebie { get; set; }
