+++ src/Core/Application/Exchange/Offers/OfferProductDetailsDto.cs
@@ -2,7 +2,7 @@ namespace FSH.WebApi.Application.Exchange.Offers;
 
 public class OfferProductDetailsDto : OfferProductDto
 {
-    public decimal NetValue { get; set; }
-    public decimal GrossValue { get; set; }
+    public long NetValue { get; set; }
+    public long GrossValue { get; set; }
     public Guid OfferId { get; set; }
 }
